Refactor: Move home components below to (main)
