+++ src/components/(main)/home/navRegisterButtons.tsx
